Update watch and serve commands
- Bundle app using spago using --watch to do it on each code changes
- Serve using esbuild also watching for changes

The docs mention that spago has -some- esbuild minimal functionality
that is enough for small projects, which this evidently is, and
recommends using esbuild as the project grows. We use both for now just
to be able to serve but according to the verbose output of spago
bundle-app we could do it in one shot. Let's go with this for now.