Change Board representation to a Map
The Bingo board is represented as a map from a letter to an array of
board values.

We make a distinction on potential board values to
existing, non-existing and not-used. The latter represents the typical
box in a board that doesn't hold any value. We add the required
functions to convert to/from textual representations to use as values
of inputs and also to display them.

We also include a function to update a value in the board based on the
letter and the index of the value in the array that represents the column.